Refactor: api 폴더 정리
