--- storix-fe/src/api/auth/artist-login.api.ts
@@ -1,25 +1,0 @@
-// src/api/auth/artist-login.api.ts
-import { apiClient } from '@/api/axios-instance'
-
-export type ArtistLoginRequest = {
-  loginId: string
-  password: string
-}
-
-export type ArtistLoginResponse = {
-  isSuccess: boolean
-  code: string
-  message: string
-  result?: {
-    accessToken?: string
-  }
-  timestamp: string
-}
-
-export async function artistLoginUser(body: ArtistLoginRequest) {
-  const res = await apiClient.post<ArtistLoginResponse>(
-    '/api/v1/auth/users/artist/login',
-    body,
-  )
-  return res.data
-}
