--- storix-fe/src/api/types.ts
@@ -1,8 +1,0 @@
-// src/api/types.ts
-export type ApiResponse<T> = {
-  isSuccess: boolean
-  code: string
-  message: string
-  result: T
-  timestamp: string
-}
