+++ storix-fe/src/app/login/page.tsx
@@ -6,7 +6,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { Splash } from '@/app/splash'
-import { getKakaoAuthUrl } from '@/api/auth/kakao.api'
+import { getKakaoAuthUrl } from '@/lib/api/auth/kakao.api'
 
 function generateNaverState() {
   if (typeof window !== 'undefined' && window.crypto?.getRandomValues) {
