+++ storix-fe/src/app/profile/settings/SettingsClient.tsx
@@ -5,8 +5,8 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import { withdrawUser } from '@/api/auth/withdraw.api'
-import { logoutUser } from '@/api/auth/logout.api'
+import { withdrawUser } from '@/lib/api/auth/withdraw.api'
+import { logoutUser } from '@/lib/api/auth/logout.api'
 import { useAuthStore } from '@/store/auth.store'
 import { useProfileStore } from '@/store/profile.store'
