+++ storix-fe/src/lib/api/profile/readerActivity.api.ts
@@ -1,5 +1,5 @@
 // src/api/profile/readerActivity.api.ts
-import { apiClient } from '@/api/axios-instance'
+import { apiClient } from '@/lib/api/axios-instance'
 
 export type SortType = 'LATEST'
