+++ storix-fe/src/lib/api/profile/readerHashtags.api.ts
@@ -1,4 +1,4 @@
-import { apiClient } from '@/api/axios-instance'
+import { apiClient } from '@/lib/api/axios-instance'
 
 export type HashtagRankingResponse = {
   rankings: Record<number, string>
