+++ storix-fe/src/lib/api/profile/readerRatings.api.ts
@@ -1,5 +1,5 @@
 // src/api/profile/readerRatings.api.ts
-import { apiClient } from '@/api/axios-instance'
+import { apiClient } from '@/lib/api/axios-instance'
 
 export type RatingCountsMap = Record<string, number>
